+++ x/leverage/keeper/grpc_query.go
@@ -91,16 +91,28 @@ func (q Querier) MarketSummary(
 	// maxBorrow is based on MaxSupplyUtilization
 	maxBorrow := token.MaxSupplyUtilization.MulInt(supplied.Amount).TruncateInt()
 
-	// availableBorrow respects maxBorrow but also liquidity
-	// TODO #1163 - additional restriction
-	availableBorrow := sdk.MinInt(liquidity, maxBorrow.Sub(borrowed.Amount))
+	// minimum liquidity respects both MaxSupplyUtilization and MinCollateralLiquidity
+	minLiquidityFromSupply := supplied.Amount.Sub(maxBorrow)
+	minLiquidityFromCollateral := token.MinCollateralLiquidity.Mul(rate.MulInt(uCollateral.Amount)).TruncateInt()
+	minLiquidity := sdk.MinInt(minLiquidityFromCollateral, minLiquidityFromSupply)
+
+	// availableBorrow respects both maxBorrow and minLiquidity
+	availableBorrow := liquidity.Sub(minLiquidity)
+	availableBorrow = sdk.MinInt(availableBorrow, maxBorrow.Sub(borrowed.Amount))
 	availableBorrow = sdk.MaxInt(availableBorrow, sdk.ZeroInt())
 
-	// minimum liquidity respects both MaxSupplyUtilization and MinCollateralLiquidity
-	// TODO #1163 - additional restriction
-	minLiquidity := supplied.Amount.Sub(maxBorrow)
+	// availableWithdraw is based on minLiquidity
+	availableWithdraw := liquidity.Sub(minLiquidity)
+	availableWithdraw = sdk.MaxInt(availableWithdraw, sdk.ZeroInt())
 
-	maxCollateral, _ := q.Keeper.maxCollateral(ctx, uDenom)
+	// availableCollateralize respects both MaxCollateralShare and MinCollateralLiquidity
+	maxCollateral, _ := q.Keeper.maxCollateralFromShare(ctx, uDenom)
+	if token.MinCollateralLiquidity.IsPositive() {
+		maxCollateralFromLiquidity := toDec(liquidity).Quo(token.MinCollateralLiquidity).TruncateInt()
+		maxCollateral = sdk.MinInt(maxCollateral, maxCollateralFromLiquidity)
+	}
+	availableCollateralize := maxCollateral.Sub(uCollateral.Amount)
+	availableCollateralize = sdk.MaxInt(availableCollateralize, sdk.ZeroInt())
 
 	resp := types.QueryMarketSummaryResponse{
 		SymbolDenom:            token.SymbolDenom,
@@ -118,8 +130,8 @@ func (q Querier) MarketSummary(
 		MinimumLiquidity:       minLiquidity,
 		UTokenSupply:           uSupply.Amount,
 		AvailableBorrow:        availableBorrow,
-		AvailableWithdraw:      uSupply.Amount, // TODO #1163 - implement limits
-		AvailableCollateralize: sdk.MaxInt(maxCollateral.Sub(uCollateral.Amount), sdk.ZeroInt()),
+		AvailableWithdraw:      availableWithdraw,
+		AvailableCollateralize: availableCollateralize,
 	}
 
 	// Oracle price in response will be nil if it is unavailable